+++ starbot-core/src/main/java/com/starlwr/bot/core/model/PushUser.java
@@ -3,10 +3,8 @@ package com.starlwr.bot.core.model;
 import com.starlwr.bot.core.enums.LivePlatform;
 import com.starlwr.bot.core.event.dynamic.StarBotBaseDynamicEvent;
 import com.starlwr.bot.core.event.live.StarBotBaseLiveEvent;
-import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.context.annotation.Profile;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,60 +13,42 @@ import java.util.Objects;
 /**
  * 推送用户
  */
-@Profile("mysql")
 @Getter
 @Setter
-@Entity
-@Table(name = "starbot_push_user")
 public class PushUser {
-    /**
-     * ID，数据库类数据源使用
-     */
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
     /**
      * UID
      */
-    @Column(name = "uid")
     private Long uid;
 
     /**
      * 昵称，非必填，会自动获取
      */
-    @Column(name = "uname")
     private String uname;
 
     /**
      * 房间号，非必填，会自动获取
      */
-    @Column(name = "room_id")
     private Long roomId;
 
     /**
      * 头像，非必填，会自动获取
      */
-    @Transient
     private String face;
 
     /**
      * 直播平台，请优先从 {@link LivePlatform} 中获取，若不存在可使用自定义字符串
      */
-    @Column(name = "platform")
     private String platform;
 
     /**
      * 是否启用
      */
-    @Column(name = "enabled")
     private Boolean enabled;
 
     /**
      * 关联的推送目标
      */
-    @OneToMany(mappedBy = "user", fetch = FetchType.EAGER, cascade = CascadeType.ALL, orphanRemoval = true)
     private List<PushTarget> targets = new ArrayList<>();
 
     @Override